// src/types/BoletosProps.ts
 export interface boletosProps {
  id: number;
  boleto_id: number;
  cliente_id: number;
  data_documento: string;
  data_pagamento: string;
  data_gerada: string;
  data_processamento: string;
  data_vencimento: string;
  email: string;
  obs_corpo: string;
  status: string;
  valor_documento: number;
  n_fantasia: string;
  numero_documento: string;
  cep: string;
  bairro: string;
  cnpj: string;
  ibge: string;
  endereco: string;
  cidade: string;
  rz_social: string;
  numero: string;
  estado:string;
 }